python-requirements.txt
django>=1.7
django-static-precompiler>=0.9
